+++ besu/src/test/java/org/hyperledger/besu/cli/BesuCommandTest.java
@@ -2015,7 +2015,7 @@ public class BesuCommandTest extends CommandTestAbstract {
   }
 
   @Test
-  public void rpcHttpHostWhitelistAcceptsSingleArgument() {
+  public void rpcHttpHostAllowlistAcceptsSingleArgument() {
     parseCommand("--host-whitelist", "a");
 
     verify(mockRunnerBuilder).jsonRpcConfiguration(jsonRpcConfigArgumentCaptor.capture());
@@ -2031,7 +2031,7 @@ public class BesuCommandTest extends CommandTestAbstract {
   }
 
   @Test
-  public void rpcHttpHostWhitelistAcceptsMultipleArguments() {
+  public void rpcHttpHostAllowlistAcceptsMultipleArguments() {
     parseCommand("--host-whitelist", "a,b");
 
     verify(mockRunnerBuilder).jsonRpcConfiguration(jsonRpcConfigArgumentCaptor.capture());
@@ -2047,7 +2047,7 @@ public class BesuCommandTest extends CommandTestAbstract {
   }
 
   @Test
-  public void rpcHttpHostWhitelistAcceptsDoubleComma() {
+  public void rpcHttpHostAllowlistAcceptsDoubleComma() {
     parseCommand("--host-whitelist", "a,,b");
 
     verify(mockRunnerBuilder).jsonRpcConfiguration(jsonRpcConfigArgumentCaptor.capture());
@@ -2063,7 +2063,7 @@ public class BesuCommandTest extends CommandTestAbstract {
   }
 
   @Test
-  public void rpcHttpHostWhitelistAcceptsMultipleFlags() {
+  public void rpcHttpHostAllowlistAcceptsMultipleFlags() {
     parseCommand("--host-whitelist=a", "--host-whitelist=b");
 
     verify(mockRunnerBuilder).jsonRpcConfiguration(jsonRpcConfigArgumentCaptor.capture());
@@ -2079,7 +2079,7 @@ public class BesuCommandTest extends CommandTestAbstract {
   }
 
   @Test
-  public void rpcHttpHostWhitelistStarWithAnotherHostnameMustFail() {
+  public void rpcHttpHostAllowlistStarWithAnotherHostnameMustFail() {
     final String[] origins = {"friend", "*"};
     parseCommand("--host-whitelist", String.join(",", origins));
 
@@ -2091,7 +2091,7 @@ public class BesuCommandTest extends CommandTestAbstract {
   }
 
   @Test
-  public void rpcHttpHostWhitelistStarWithAnotherHostnameMustFailStarFirst() {
+  public void rpcHttpHostAllowlistStarWithAnotherHostnameMustFailStarFirst() {
     final String[] origins = {"*", "friend"};
     parseCommand("--host-whitelist", String.join(",", origins));
 
@@ -2103,7 +2103,7 @@ public class BesuCommandTest extends CommandTestAbstract {
   }
 
   @Test
-  public void rpcHttpHostWhitelistAllWithAnotherHostnameMustFail() {
+  public void rpcHttpHostAllowlistAllWithAnotherHostnameMustFail() {
     final String[] origins = {"friend", "all"};
     parseCommand("--host-whitelist", String.join(",", origins));
 
@@ -2115,7 +2115,7 @@ public class BesuCommandTest extends CommandTestAbstract {
   }
 
   @Test
-  public void rpcHttpHostWhitelistWithNoneMustBuildEmptyList() {
+  public void rpcHttpHostAllowlistWithNoneMustBuildEmptyList() {
     final String[] origins = {"none"};
     parseCommand("--host-whitelist", String.join(",", origins));
 
@@ -2129,7 +2129,7 @@ public class BesuCommandTest extends CommandTestAbstract {
   }
 
   @Test
-  public void rpcHttpHostWhitelistNoneWithAnotherDomainMustFail() {
+  public void rpcHttpHostAllowlistNoneWithAnotherDomainMustFail() {
     final String[] origins = {"http://domain1.com", "none"};
     parseCommand("--host-whitelist", String.join(",", origins));
 
@@ -2141,7 +2141,7 @@ public class BesuCommandTest extends CommandTestAbstract {
   }
 
   @Test
-  public void rpcHttpHostWhitelistNoneWithAnotherDomainMustFailNoneFirst() {
+  public void rpcHttpHostAllowlistNoneWithAnotherDomainMustFailNoneFirst() {
     final String[] origins = {"none", "http://domain1.com"};
     parseCommand("--host-whitelist", String.join(",", origins));
 
@@ -2153,7 +2153,7 @@ public class BesuCommandTest extends CommandTestAbstract {
   }
 
   @Test
-  public void rpcHttpHostWhitelistEmptyValueFails() {
+  public void rpcHttpHostAllowlistEmptyValueFails() {
     parseCommand("--host-whitelist=");
 
     Mockito.verifyZeroInteractions(mockRunnerBuilder);
@@ -3080,7 +3080,7 @@ public class BesuCommandTest extends CommandTestAbstract {
         staticNodesFile.toPath(), ("[\"" + staticNodeURI.toString() + "\"]").getBytes(UTF_8));
     Files.write(
         permissioningConfig.toPath(),
-        ("nodes-whitelist=[\"" + allowedNode.toString() + "\"]").getBytes(UTF_8));
+        ("nodes-allowlist=[\"" + allowedNode.toString() + "\"]").getBytes(UTF_8));
 
     parseCommand(
         "--data-path=" + testFolder.getRoot().getPath(),
@@ -3088,7 +3088,7 @@ public class BesuCommandTest extends CommandTestAbstract {
         "--permissions-nodes-config-file-enabled=true",
         "--permissions-nodes-config-file=" + permissioningConfig.getPath());
     assertThat(commandErrorOutput.toString())
-        .contains(staticNodeURI.toString(), "not in nodes-whitelist");
+        .contains(staticNodeURI.toString(), "not in nodes-allowlist");
   }
 
   @Test