+++ ethereum/permissioning/src/main/java/org/hyperledger/besu/ethereum/permissioning/PermissioningConfigurationBuilder.java
@@ -26,8 +26,10 @@ import org.apache.tuweni.toml.TomlParseResult;
 
 public class PermissioningConfigurationBuilder {
 
-  public static final String ACCOUNTS_WHITELIST_KEY = "accounts-whitelist";
-  public static final String NODES_WHITELIST_KEY = "nodes-whitelist";
+  @Deprecated public static final String ACCOUNTS_WHITELIST_KEY = "accounts-whitelist";
+  @Deprecated public static final String NODES_WHITELIST_KEY = "nodes-whitelist";
+  public static final String ACCOUNTS_ALLOWLIST_KEY = "accounts-allowlist";
+  public static final String NODES_ALLOWLIST_KEY = "nodes-allowlist";
 
   public static SmartContractPermissioningConfiguration smartContractPermissioningConfiguration(
       final Address address, final boolean smartContractPermissionedNodeEnabled) {
@@ -65,7 +67,8 @@ public class PermissioningConfigurationBuilder {
 
     if (localConfigNodePermissioningEnabled) {
       final TomlParseResult nodePermissioningToml = readToml(nodePermissioningConfigFilepath);
-      final TomlArray nodeWhitelistTomlArray = nodePermissioningToml.getArray(NODES_WHITELIST_KEY);
+      final TomlArray nodeWhitelistTomlArray =
+          getAllowlistArray(nodePermissioningToml, NODES_ALLOWLIST_KEY, NODES_WHITELIST_KEY);
 
       permissioningConfiguration.setNodePermissioningConfigFilePath(
           nodePermissioningConfigFilepath);
@@ -81,7 +84,7 @@ public class PermissioningConfigurationBuilder {
         permissioningConfiguration.setNodeAllowlist(nodesWhitelistToml);
       } else {
         throw new Exception(
-            NODES_WHITELIST_KEY
+            NODES_ALLOWLIST_KEY
                 + " config option missing in TOML config file "
                 + nodePermissioningConfigFilepath);
       }
@@ -98,7 +101,8 @@ public class PermissioningConfigurationBuilder {
     if (localConfigAccountPermissioningEnabled) {
       final TomlParseResult accountPermissioningToml = readToml(accountPermissioningConfigFilepath);
       final TomlArray accountWhitelistTomlArray =
-          accountPermissioningToml.getArray(ACCOUNTS_WHITELIST_KEY);
+          getAllowlistArray(
+              accountPermissioningToml, ACCOUNTS_ALLOWLIST_KEY, ACCOUNTS_WHITELIST_KEY);
 
       permissioningConfiguration.setAccountPermissioningConfigFilePath(
           accountPermissioningConfigFilepath);
@@ -122,13 +126,32 @@ public class PermissioningConfigurationBuilder {
         permissioningConfiguration.setAccountAllowlist(accountsWhitelistToml);
       } else {
         throw new Exception(
-            ACCOUNTS_WHITELIST_KEY
+            ACCOUNTS_ALLOWLIST_KEY
                 + " config option missing in TOML config file "
                 + accountPermissioningConfigFilepath);
       }
     }
 
     return permissioningConfiguration;
+  }
+
+  /**
+   * This method allows support for both keys for now. Whitelist TOML keys will be removed in future
+   * (breaking change)
+   *
+   * @param tomlParseResult
+   * @param primaryKey
+   * @param alternateKey
+   * @return
+   */
+  private static TomlArray getAllowlistArray(
+      final TomlParseResult tomlParseResult, final String primaryKey, final String alternateKey) {
+    final TomlArray array = tomlParseResult.getArray(primaryKey);
+    if (array == null) {
+      return tomlParseResult.getArray(alternateKey);
+    } else {
+      return array;
+    }
   }
 
   private static TomlParseResult readToml(final String filepath) throws Exception {